% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{chicago_air}
\alias{chicago_air}
\title{Air quality data from a Chicago monitor.}
\format{
A data frame with 730 rows and 6 variables:
\describe{
  \item{date}{Year-Month-Day}
  \item{ozone}{8-hour maximum value in ppm}
  \item{temp}{1-hour maximum value in degrees Fahrenheit}
  \item{pressure}{1-hour maximum value in Millibars}
  \item{month}{Month (1-12)}
  \item{weekday}{Day of week (1-7)}
}
}
\source{
\url{https://aqs.epa.gov/aqsweb/airdata/download_files.html}
}
\usage{
chicago_air
}
\description{
A dataset containing daily values of ozone, temperature, and barometric pressure from a Chicago monitor between
January 1, 2021 and December 31, 2021.
}
\details{
These are measurements from a monitor with EPA AQS code 17023001 at latitude 39.21086 and longitude -87.6683,
datum WGS84.
}
\keyword{datasets}
