% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{chicago_aqs}
\alias{chicago_aqs}
\title{Air quality data from AQS}
\format{
A data frame with 1009 rows and 34 columns:
\describe{
  \item{State_Code}{Code for the state where the monitor is located}
  \item{County_Code}{Code for the county where the monitor is located}
  \item{Site_Number}{Numer identifying a particular site in a county}
  \item{Parameter_Code}{Code identifying what is being measured}
  \item{POC}{“Parameter Occurrence Code” used to distinguish different instruments that measure the same parameter at the same site}
  \item{Latitude}{Latitude of the monitor location}
  \item{Longitude}{Longitude of the monitor location}
  \item{Datum}{The Datum associated with the Latitude and Longitude measures}
  \item{Parameter_Name}{Name of the parameter}
  \item{Duration_Description}{the length of time used to acquire a sample measurement}
  \item{Pollutant_Standard}{The Air Quality Standard applicable to the measurement}
  \item{Date_Local}{Date of the measurement}
  \item{Year}{Year of the measurement}
  \item{Day_In_Year_Local}{Day of the year}
  \item{Units_of_Measure}{Unit}
  \item{Exceptional_Data_Type}{Flag for exceptional events}
  \item{Nonreg_Observation_Count}
  \item{Ovservation_Count}
  \item{Observation_Percent}
  \item{Nonreg_Arithmetic_Mean}
  \item{Arithmetic_Mean}
  \item{Nonreg_First_Maximum_Value}
  \item{First_Maximum_Value}
  \item{First_Maximum_Hour}
  \item{AQI}
  \item{Daily_Criteria_Indicator}
  \item{Tribe_Name}
  \item{State_Name}
  \item{County_Name}
  \item{City_Name}
  \item{Local_Site_Name}
  \item{Address}
  \item{MSA_or_CBSA_Name}
  \item{Data_Source}
}
}
\source{
\url{https://aqs.epa.gov/aqsweb/airdata/download_files.html}
}
\usage{
chicago_aqs
}
\description{
Air quality data from AQS
}
\keyword{datasets}
