% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{chicago_wind}
\alias{chicago_wind}
\title{Wind data from a Chicago monitor.}
\format{A data frame with 8643 rows and 4 variables:
\describe{
  \item{datetime}{Date and time of the data value, given in the following format without spaces: YYYYMMDDThhmmTZD. Four-digit year (YYYY), two-digit month (MM), two-digit day (DD), capital letter T (time), two-digit hour (hh, in 24-hr time), two-digit minutes (mm), and the time zone designation, TZD. (The date-time field is the ISO 8601 Basic.) TZD is +hhmm/-hhmm from GMT. The time corresponds to the begin time of the sampling period.}
  \item{wind_speed}{1-hour resultant value in knots}
  \item{wind_dir}{1-hour resultant wind direction in degrees compass}
  \item{ozone}{1-hour value in ppm}
}}
\source{
\url{https://ofmext.epa.gov/AQDMRS/aqdmrs.html}
}
\usage{
chicago_wind
}
\description{
A dataset containing hourly values of wind speed, wind direction, and ozone concentration from a Chicago monitor between
January 1, 2013 and December 31, 2013.
}
\details{
These are measurements from a monitor with EPA AQS code 840170310001 at latitude 41.67099 and longitude -87.73246,
datum WGS84.
}
\keyword{datasets}

