% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ertac_egu_projections}
\alias{ertac_egu_projections}
\title{ERTAC EGU Air Emissions Projections}
\format{
A data frame with 80,551 rows and 18 variables:
\describe{
  \item{oris_code}{(integer) ORIS Plant Code, a unique identifier for power plants in the United States.}
  \item{unit_id}{(character) Identifier for each unit within a plant.}
  \item{facility_name}{(character) Name of the power generation facility.}
  \item{state}{(character) US state where the facility is located.}
  \item{fips_code}{(character) Federal Information Processing Standard code for the county.}
  \item{ertac_region}{(character) ERTAC region code indicating the regional grouping for analysis.}
  \item{fuel_type}{(character) Type of fuel used for power generation (e.g., Coal, Gas).}
  \item{gen_capacity_mw}{(numeric) Generation capacity of the unit in megawatts.}
  \item{annual_so2_2022_tons}{(numeric) Projected annual SO2 emissions in tons for the year 2022.}
  \item{avg_annual_so2_rate_2022}{(numeric) Projected average annual SO2 emissions rate in pounds per million British thermal units (lbs/MMBtu) for 2022.}
  \item{annual_nox_2022_tons}{(numeric) Projected annual NOx emissions in tons for the year 2022.}
  \item{avg_annual_nox_rate_2022}{(numeric) Projected average annual NOx emissions rate in pounds per million British thermal units (lbs/MMBtu) for 2022.}
  \item{annual_so2_2038_tons}{(numeric) Projected annual SO2 emissions in tons for the year 2038.}
  \item{avg_annual_so2_rate_2038}{(numeric) Projected average annual SO2 emissions rate in pounds per million British thermal units (lbs/MMBtu) for 2038.}
  \item{annual_nox_2038_tons}{(numeric) Projected annual NOx emissions in tons for the year 2038.}
  \item{avg_annual_nox_rate_2038}{(numeric) Projected average annual NOx emissions rate in pounds per million British thermal units (lbs/MMBtu) for 2038.}
  \item{longitude}{(numeric) Longitude of the facility location.}
  \item{latitude}{(numeric) Latitude of the facility location.}
}
}
\source{
\url{https://www.epa.gov/air-emissions-inventories/eastern-regional-technical-advisory-committee-ertac-electricity}
}
\usage{
ertac_egu_projections
}
\description{
A dataset containing projected air emissions data for Electric Generating Units (EGUs) within the ERTAC region, 
including details such as facility name, location, fuel type, generation capacity, and projected annual emissions 
rates for SO2 and NOx for the years 2022 and 2038.
}
\details{
The dataset is intended for use by researchers and policymakers to analyze future emissions scenarios and 
to plan for regulatory compliance and environmental impact assessments.
}
\keyword{datasets}
