% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{chicago_air}
\alias{chicago_air}
\title{Air Quality Data from a Chicago monitor.}
\format{A data frame with 730 rows and 6 variables:
\describe{
  \item{date}{Year-Month-Day}
  \item{ozone}{8-hour maximum value in ppm}
  \item{temp}{1-hour maximum value in degrees Fahrenheit}
  \item{solar}{1-hour maximum value in Langleys/minute}
  \item{month}{Month (1-12)}
  \item{weekday}{Day of week (1-7)}
}}
\source{
\url{https://ofmext.epa.gov/AQDMRS/aqdmrs.html}
}
\usage{
chicago_air
}
\description{
A dataset containing daily values of ozone, temperature, and solar radiation from a Chicago monitor between
January 1, 2013 and December 31, 2013.
}
\details{
These are measurements from a monitor with EPA AQS code 170314201 at latitude 42.14 and longitude -87.79923.
}
\keyword{datasets}

